+++ publish/CHANGELOG.md
@@ -1,3 +1,11 @@
+- v1.40
+  - Adds a new command `hammer_restore` that selects a blueprint and automatically sets hammer at its position and rotation.
+  - Adds a new command `hammer_pos` that allows setting the hammer position and rotation.
+  - Changes the `hammer_save` command to save hammer position and rotation to the blueprint file.
+  - Changes the `hammer_save` command to print hammer position and rotation to the console.
+  - Changes the `hammer_save` command to automatically load the saved blueprint.
+  - Changes the `hammer_grid` to use X,Z,Y format instead of X,Y,Z.
+
 - v1.39
   - Fixed for the new patch.